fix: Fix an NPE when deleting a nonexistent registry (#467)

+++ backend/sdk/src/main/java/com/alibaba/higress/sdk/service/ServiceSourceServiceImpl.java
@@ -125,6 +125,10 @@ class ServiceSourceServiceImpl implements ServiceSourceService {
             return;
         }
         V1RegistryConfig registry = kubernetesModelConverter.removeV1McpBridgeRegistry(mcpBridge, name);
+        if (registry == null) {
+            // There is nothing to delete.
+            return;
+        }
         if (StringUtils.isNotEmpty(registry.getAuthSecretName())) {
             try {
                 kubernetesClientService.deleteSecret(registry.getAuthSecretName());
